+++ Mechanical_workshop/DTOs/EstimateFullDto.cs
@@ -46,7 +46,7 @@ namespace Mechanical_workshop.Dtos
         public int ID { get; set; }
         public string Type { get; set; } = "[LABOR]";
         public string Description { get; set; } = string.Empty;
-        public int Duration { get; set; }
+        public decimal Duration { get; set; }
         public decimal LaborRate { get; set; }
         public decimal ExtendedPrice { get; set; }
         public bool Taxable { get; set; }
